requirements.txt
boto3==1.12.38
botocore==1.15.38
docutils==0.15.2
jmespath==0.9.5
Pillow==7.1.1
pkg-resources==0.0.0
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.8
